+++ src/Faker/Container/ContainerBuilder.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Faker\Container;
 
-use Faker\English\FileExtension;
+use Faker\Core\FileExtension;
 use Faker\Extension\File;
 use Psr\Container\ContainerInterface;
